fix: подстраивает приложения под критерии
- сортируем элементы после обновления правильно
- переносим константы с сообщениями пользователю в отдельный файл const.js
- меняет название переменной SortView на TripSortView (опираемся на название файлов)
- чистим редактор новой точки маршрута (даты не должны отображаться в том числе)
- исправляет ошибку с выделением всех предложений при смене типа точки маршрута
