+++ src/const.js
@@ -7,6 +7,15 @@ const FilterType = {
   PAST: 'past',
 };
 
+const EmptyMessage = {
+  [FilterType.EVERYTHING]: 'Click New Event to create your first point',
+  [FilterType.FUTURE]: 'There are no future events now',
+  [FilterType.PAST]: 'There are no past events now',
+  [FilterType.PRESENT]: 'There are no present events now',
+  LOADING: 'Loading...',
+  FAILURE: 'Failed to load latest route information',
+};
+
 const SortType = {
   DAY: 'day',
   EVENT: 'event',
@@ -29,4 +38,4 @@ const UpdateType = {
   FAILURE: 'FAILURE',
 };
 
-export { POINT_TYPES, FilterType, SortType, UserAction, UpdateType };
+export { POINT_TYPES, FilterType, EmptyMessage, SortType, UserAction, UpdateType };
