+++ src/view/message-view.js
@@ -1,14 +1,5 @@
 import AbstractView from '../framework/view/abstract-view';
-import { FilterType } from '../const';
-
-const EmptyMessage = {
-  [FilterType.EVERYTHING]: 'Click New Event to create your first point',
-  [FilterType.FUTURE]: 'There are no future events now',
-  [FilterType.PAST]: 'There are no past events now',
-  [FilterType.PRESENT]: 'There are no present events now',
-  LOADING: 'Loading...',
-  FAILURE: 'Failed to load latest route information',
-};
+import { EmptyMessage } from '../const';
 
 function createMessageTemplate(messageText) {
   const emptyMessageValue = EmptyMessage[messageText];
